organize header, create custom exception class
