+++ python/pybind11_bindings/include/openravepy/openravepy_manipulatorinfo.h
